Refactor: 패키지 리팩토링
패키지를 용도에 따라 리팩토링함
