+++ src/main/java/car/domain/winnerstrategy/MaxPositionDuplicateWinnerStrategy.java
@@ -1,6 +1,6 @@
-package winnerstrategy;
+package car.domain.winnerstrategy;
 
-import car.Car;
+import car.domain.Car;
 import java.util.List;
 import java.util.stream.Collectors;
