+++ src/test/java/car/domain/winnerstrategy/MaxPositionDuplicateWinnerStrategyTest.java
@@ -1,8 +1,10 @@
-package winnerstrategy;
+package car.domain.winnerstrategy;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import car.Car;
+import car.domain.Car;
+import car.domain.winnerstrategy.MaxPositionDuplicateWinnerStrategy;
+import car.domain.winnerstrategy.WinnerStrategy;
 import java.util.List;
 import java.util.stream.Collectors;
 import org.junit.jupiter.api.DisplayName;
